Resolve "Being able to delete your own submissions (as user)"

See merge request karaokemugen/karaokemugen-app!155

+++ src/_services/engine.js
@@ -617,7 +617,9 @@ export async function deletePL(playlist_id, token) {
 	}
 }
 
-export async function deleteKara(plc_ids,playlist_id) {
+export async function deleteKara(plc_ids,playlist_id,token) {
+	// If playlist_id is null, set it to current/public PL ID
+	if (!playlist_id) playlist_id = internalState.modePlaylistID;
 	const pl = await plc.getPlaylistInfo(playlist_id);
 	let karas;
 	if (typeof plc_ids === 'string') {
@@ -628,6 +630,8 @@ export async function deleteKara(plc_ids,playlist_id) {
 	const plcData = await plc.getPLCInfoMini(karas[0]);
 	logger.info(`[Engine] Deleting karaokes from playlist ${pl.name} : ${plcData.title}...`);	
 	try {
+		//If token is present, a user is trying to remove a karaoke
+		if (token) if (plcData.username !== token.username) throw 'You cannot delete a song you did not add';
 		await plc.deleteKaraFromPlaylist(karas, playlist_id);
 		return {
 			pl_id: playlist_id,
